add input validation

+++ api/Controllers/StockController.cs
@@ -25,7 +25,7 @@ namespace api.Controllers
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stockDto);
         }
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetStockById([FromRoute] int id)
         {
             var stock = await _stockRepo.GetByIdAsync(id);
@@ -35,22 +35,25 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var stockModel = stockDto.ToStockFromCreateDto();
             await _stockRepo.CreatAysnc(stockModel);
             return CreatedAtAction(nameof(GetStockById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var stockModel = await _stockRepo.UpdateAsync(id, stockDto);
             if (stockModel == null) return NotFound();
 
             return Ok(stockModel.ToStockDto());
         }
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var stockModel = await _stockRepo.DeleteAsync(id);
             if (stockModel == null) return NotFound();
             return NoContent();
